+++ internal/log/log.go
@@ -7,7 +7,6 @@ import (
 	"path/filepath"
 
 	"github.com/sirupsen/logrus"
-	"google.golang.org/grpc/codes"
 )
 
 // NewStdout returns a logger printing its results to stdout.
@@ -114,46 +113,4 @@ func LevelFrom(level int) logrus.Level {
 		return logrus.TraceLevel
 	}
 	return logrus.DebugLevel
-}
-
-// Maps a grpc response code to a logging level
-func CodeToLevel(code codes.Code) logrus.Level {
-	switch code {
-	case codes.OK:
-		return logrus.DebugLevel
-	case codes.Canceled:
-		return logrus.DebugLevel
-	case codes.Unknown:
-		return logrus.ErrorLevel
-	case codes.InvalidArgument:
-		return logrus.WarnLevel
-	case codes.DeadlineExceeded:
-		return logrus.WarnLevel
-	case codes.NotFound:
-		return logrus.DebugLevel
-	case codes.AlreadyExists:
-		return logrus.DebugLevel
-	case codes.PermissionDenied:
-		return logrus.WarnLevel
-	case codes.Unauthenticated:
-		return logrus.WarnLevel
-	case codes.ResourceExhausted:
-		return logrus.WarnLevel
-	case codes.FailedPrecondition:
-		return logrus.WarnLevel
-	case codes.Aborted:
-		return logrus.WarnLevel
-	case codes.OutOfRange:
-		return logrus.WarnLevel
-	case codes.Unimplemented:
-		return logrus.ErrorLevel
-	case codes.Internal:
-		return logrus.ErrorLevel
-	case codes.Unavailable:
-		return logrus.WarnLevel
-	case codes.DataLoss:
-		return logrus.ErrorLevel
-	default:
-		return logrus.ErrorLevel
-	}
 }
